server/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
pydantic
python-multipart
python-dotenv
httpx
google-cloud-translate
pillow
